% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling_r_functions.R
\name{PrepareNonPara}
\alias{PrepareNonPara}
\title{Prepares a list for non-parametric testing}
\usage{
PrepareNonPara(X, Y, permutation.num, iter.num, samp.size)
}
\arguments{
\item{X}{A matrix of nx observations on p dimensions}

\item{Y}{A matrix of ny observations on p dimensions}

\item{iter.num}{Number of samples of dimensions to conduct}

\item{samp.size}{Number of dimensions to sample at each sampling}

\item{permutation.mat}{Number of permutations to conduct}
}
\value{
A list which consists of Original X and Y matrices, Permutation a list length of permutation.num
consists of shuffled X and Y and samp.mat a matrix of iter.num on samp.size dimensions
}
\description{
Prepares a list for non-parametric testing
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(20, rep(0, 50), diag(50)) 
Y <- mvrnorm(20, rep(0, 50), diag(50))
L <- PrepareNonPara(X, Y,
permutation.num = 100, 
iter.num = 100, 
samp.size = 50) 
}
}
