% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling R Functions.R
\name{SHnonPara}
\alias{SHnonPara}
\title{Conducts non parametric test, based on sampling of dimensions}
\usage{
SHnonPara(permute.list, method = "SH")
}
\arguments{
\item{permute.list}{a list output from PrepareNonPara}

\item{method}{the type of test to conduct SH or Thullin}
}
\value{
A list containing the original statistic defined by method, permutation statistic vector, 
the p-value of the permutaiton test, method used in the test and number of permutations
}
\description{
Conducts non parametric test, based on sampling of dimensions
}
