% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling R Functions.R
\name{SHTest}
\alias{SHTest}
\title{Conduct Simes Hotelling test}
\usage{
SHTest(X, Y = NULL, samp.size = NULL, iterations = nrow(X) *
  log(nrow(X)), equal.cov = TRUE, delta = rep(0, ncol(X)))
}
\arguments{
\item{X}{A matrix of nx observations on p dimensions}

\item{Y}{A matrix of ny observations on p dimensions}

\item{samp.size}{Number of dimensions (m) sampled at each iteration}

\item{iterations}{Number of iterations (B) conducted}

\item{equal.cov}{Logical, indicating if the equal.cov version of the test should be conducted}

\item{delta}{Vector of expected differences between the mean vector of X and Y under the null}
}
\value{
A list including the dimensions used for analysis, the simes P-value, and the parameters used to conduct the test, including samp.size, iteration and the equal.cov flag.
}
\description{
Conduct Simes Hotelling test
}
