% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling R Functions.R
\name{simesTest}
\alias{simesTest}
\title{Conduct Simes test}
\usage{
simesTest(p.vec)
}
\arguments{
\item{p.vec}{A vector of p-values}
}
\value{
The simes p-value
}
\description{
Conduct Simes test
}
