% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling_r_functions.R
\name{SHnonPara}
\alias{SHnonPara}
\title{Conducts non parametric test, based on sampling of dimensions}
\usage{
SHnonPara(permute.list, method = "SH")
}
\arguments{
\item{permute.list}{a list output from PrepareNonPara}

\item{method}{the type of test to conduct SH or Thullin}
}
\value{
A list containing the original statistic defined by method, permutation statistic vector, 
the p-value of the permutaiton test, method used in the test and number of permutations
}
\description{
Conducts non parametric test, based on sampling of dimensions
}
\examples{
\dontrun{
library(MASS)
X <- mvrnorm(20, rep(0, 50), diag(50)) 
Y <- mvrnorm(20, rep(0, 50), diag(50))
L <- PrepareNonPara(X, Y,
permutation.num = 100, 
iter.num = 100, 
samp.size = 50) 
SHnonPara(L, method = 'SH')
SHnonPara(L, method = 'Thulin')
}
}
