% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimesHoteling_r_functions.R
\name{SHTest}
\alias{SHTest}
\title{Conduct Simes Hotelling test}
\usage{
SHTest(
  X,
  Y = NULL,
  samp.size = NULL,
  iterations = nrow(X) * log(nrow(X)),
  equal.cov = TRUE,
  delta = rep(0, ncol(X))
)
}
\arguments{
\item{X}{A matrix of nx observations on p dimensions}

\item{Y}{A matrix of ny observations on p dimensions}

\item{samp.size}{Number of dimensions (m) sampled at each iteration}

\item{iterations}{Number of iterations (B) conducted}

\item{equal.cov}{Logical, indicating if the equal.cov version of the test should be conducted}

\item{delta}{Vector of expected differences between the mean vector of X and Y under the null}
}
\value{
A list including the dimensions used for analysis, the simes P-value, and the parameters used to conduct the test, including samp.size, iteration and the equal.cov flag.
}
\description{
Conduct Simes Hotelling test
}
\examples{
library(MASS)
X <- mvrnorm(50, rep(0, 100), diag(100)) 
### One-sample global null test 
SHTest(X = X,
      samp.size = 30,
      iterations = 100,
      delta = rep(0, 100))
\dontrun{ 

Y <- mvrnorm(50, rep(0, 100), diag(100)) 
### Two-sample global null test 
SHTest(X = X,
      Y = Y, 
      samp.size = 30,
      iterations = 100,
      delta = rep(0, 100))
### Non-equal covariance two-sample global null test        
### Generating data 
X <- mvrnorm(50, rep(0, 100), diag(100)) 
Y <- mvrnorm(79, rep(0, 100), 2 * diag(100))
SHTest(X, 
      Y, 
      samp.size = 40, 
      iterations = 100, 
      equal.cov = FALSE,
      delta = rep(0, 100))
}
}
