\name{simesHotelling-package}
\alias{simesHotelling-package}
\alias{simesHotelling}
\docType{package}
\title{
\packageTitle{simesHotelling}
}
\description{
\packageDescription{simesHotelling}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{simesHotelling}
\packageIndices{simesHotelling}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{simesHotelling}

Maintainer: \packageMaintainer{simesHotelling}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
